-some 64bits fixes (patch by smogm thx)

+++ branches/nhd2-exp/src/driver/netfile.cpp
@@ -902,7 +902,7 @@ FILE *f_open(const char *filename, const char *acctype)
 								return NULL;
 							}
 
-							dprintf(stderr, "f_open: adding stream %x to cache[%d]\n", (int) fd, i);
+							dprintf(stderr, "f_open: adding stream %x to cache[%d]\n", (size_t) fd, i);
 
 							cache[i].fd       = fd;
 							cache[i].csize    = CACHESIZE;
@@ -1156,7 +1156,7 @@ int f_close(FILE *stream)
 
 	if(cache[i].fd == stream)
 	{
-		dprintf(stderr, "f_close: removing stream %x from cache[%d]\n", (int)stream, i);
+		dprintf(stderr, "f_close: removing stream %x from cache[%d]\n", (size_t)stream, i);
 
 		cache[i].closed = 1;		/* indicate that the cache is closed */
 
@@ -1289,7 +1289,8 @@ char *f_type(FILE *stream, char *type)
 	/* if the stream could not be found, look for a free slot ... */
 	if(i == CACHEENTMAX)
 	{
-	dprintf(stderr, "stream %x not in type table, ", (int) stream);
+	
+	dprintf(stderr, "stream %x not in type table, ", (size_t) stream);
 
 	for(i=0 ; (i<CACHEENTMAX) && (stream_type[i].stream != NULL); i++){};
 
@@ -1300,7 +1301,7 @@ char *f_type(FILE *stream, char *type)
 		{
 			stream_type[i].stream = stream;
 			strncpy(stream_type[i].type, type, 64);
-			dprintf(stderr, "added entry (%s) for %x\n", type, (int) stream);
+			dprintf(stderr, "added entry (%s) for %x\n", type, (size_t) stream);
 		}
 		return type;
 	}
@@ -1312,7 +1313,7 @@ char *f_type(FILE *stream, char *type)
 	/* the stream is already in the table */
 	else
 	{
-		dprintf(stderr, "stream %x lookup in type table succeded\n", (int) stream);
+		dprintf(stderr, "stream %x lookup in type table succeded\n", (size_t) stream);
 
 		if(!type)
 			return stream_type[i].type;
@@ -1375,8 +1376,9 @@ int push(FILE *fd, char *buf, long len)
 
 	//	dprintf(stderr, "push: %d bytes to store [filled: %d of %d], stream: %x\n", len, cache[i].filled, CACHESIZE, fd);
 
-	if(cache[i].fd != fd) {
-		dprintf(stderr, "push: no cache present for stream %0x\n", (int) fd);
+	if(cache[i].fd != fd) 
+	{
+		dprintf(stderr, "push: no cache present for stream %0x\n", (size_t) fd);
 		return -1;
 	}
 	do
@@ -1455,7 +1457,7 @@ int push(FILE *fd, char *buf, long len)
 
 	} while(rval < len);
 
-	dprintf(stderr, "push: exitstate: [filled: %3.1f %%], stream: %x\r", 100.0 * (float)cache[i].filled / (float)cache[i].csize, (int) fd);
+	dprintf(stderr, "push: exitstate: [filled: %3.1f %%], stream: %x\r", 100.0 * (float)cache[i].filled / (float)cache[i].csize, (size_t) fd);
 
 	return rval;
 }
@@ -1470,8 +1472,7 @@ int pop(FILE *fd, char *buf, long len)
 	if(i < 0)
 		return -1;
 
-	dprintf(stderr, "pop: %d bytes requested [filled: %d of %d], stream: %x buf 0x%x\n",
-		(int) len, (int) cache[i].filled, (int) CACHESIZE, (int)fd, (int) buf);
+	dprintf(stderr, "pop: %d bytes requested [filled: %d of %d], stream: %x buf 0x%x\n", (int) len, (int) cache[i].filled, (int) CACHESIZE, (size_t)fd, (size_t) buf);
 
 	if(cache[i].fd == fd)
 	{
@@ -1525,8 +1526,7 @@ int pop(FILE *fd, char *buf, long len)
 
 					if(amt[j])
 					{
-						dprintf(stderr, "pop(): rptr: 0x%08x, buf: 0x%08x, amt[%d]=%d, blen=%d, len=%d, rval=%d\n",
-							(int) cache[i].rptr, (int) buf, j, amt[j], blen, (int) len, rval);
+						dprintf(stderr, "pop(): rptr: 0x%08x, buf: 0x%08x, amt[%d]=%d, blen=%d, len=%d, rval=%d\n", (size_t) cache[i].rptr, (size_t) buf, j, amt[j], blen, (int) len, rval);
 
 						memmove(buf, cache[i].rptr, amt[j]);
 
@@ -1540,7 +1540,7 @@ int pop(FILE *fd, char *buf, long len)
 					}
 				}
 
-				dprintf(stderr, "pop: %d/%d/%d bytes read [filled: %d of %d], stream: %x\n", amt[0] + amt[1], rval, (int) len, (int) cache[i].filled, (int) CACHESIZE, (int) fd);
+				dprintf(stderr, "pop: %d/%d/%d bytes read [filled: %d of %d], stream: %x\n", amt[0] + amt[1], rval, (int) len, (int) cache[i].filled, (int) CACHESIZE, (size_t) fd);
 
 				/* if the cache is closed and empty, then */
 				/* force the end condition to be met */
@@ -1569,7 +1569,7 @@ int pop(FILE *fd, char *buf, long len)
 	}
 	else
 	{
-		dprintf(stderr, "pop: no cache present for stream %0x\n", (int) fd);
+		dprintf(stderr, "pop: no cache present for stream %0x\n", (size_t) fd);
 		rval = -1;
 	}
 
@@ -1593,7 +1593,7 @@ void CacheFillThread(void *c)
 	if(scache->closed)
 		return;
 
-	dprintf(stderr, "CacheFillThread: thread started, using stream %8x\n", (int) scache->fd);
+	dprintf(stderr, "CacheFillThread: thread started, using stream %8x\n", (size_t) scache->fd);
 
 	buf = (char*)malloc(CACHEBTRANS);
 
@@ -1651,7 +1651,7 @@ void CacheFillThread(void *c)
 	pthread_mutex_unlock( &scache->readable );
 
 	/* ... and exit this thread. */
-	dprintf(stderr, "CacheFillThread: thread exited, stream %8x  \n", (int) scache->fd);
+	dprintf(stderr, "CacheFillThread: thread exited, stream %8x  \n", (size_t) scache->fd);
 
 	free(buf);
 	pthread_exit(0);
@@ -1703,7 +1703,7 @@ void ShoutCAST_ParseMetaData(char *md, CSTATE *state)
 	if((!md) || (!state))
 		return;
 
-	dprintf(stderr, "ShoutCAST_ParseMetaData(%x : %s, %x)\n", (int) md, md, (int) state);
+	dprintf(stderr, "ShoutCAST_ParseMetaData(%x : %s, %x)\n", (size_t) md, md, (size_t) state);
 
 	ptr = strstr(md, "StreamTitle=");
